# lib/api-zod/src/generated/types/createFoodLogRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kitchen CFO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFoodLogRequestMealType } from "./createFoodLogRequestMealType";
import type { NutrientData } from "./nutrientData";

export interface CreateFoodLogRequest {
  foodName: string;
  quantity?: string;
  mealType: CreateFoodLogRequestMealType;
  nutrients: NutrientData;
  imageBase64?: string | null;
  notes?: string | null;
  loggedAt?: Date | null;
}